# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Attendance & Classroom Verification API (ATT)
  version: 1.1.0
  description: |
    Backend API for the Smart Attendance project (requirement prefix ATT).

    **Authentication:** call `POST /auth/login`, then send the returned token in every request as
    `Authorization: Bearer <token>`. Tokens are valid for 8 hours.

    **Roles:** `student`, `lecturer`, `ta`, `admin`, `auditor`. Authorization is enforced on the backend.

    **QR tokens:** the lecturer/TA app calls `GET /sessions/{id}/qr` repeatedly (about every 15 seconds).
    Each token expires after 20 seconds. The student app sends the scanned token to `POST /attendance/scan`.

    **Dates and times:** dates are `YYYY-MM-DD`, timestamps are ISO 8601 (UTC), times are `HH:MM:SS`.

    **Rate limiting:** `POST /attendance/scan` allows 10 attempts per minute per student account
    (HTTP 429 with `reason: too_many_attempts` afterwards).

    **Flags:** `GET /flags` returns rule-based, advisory-only attendance flags. They never change
    attendance and can be unavailable (HTTP 503) without affecting attendance capture.

    All data used in this project is synthetic.
servers:
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /me:
    get:
      tags: [Auth]
      summary: Current user (from the token)
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  role:
                    $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /attendance/scan:
    post:
      tags: [Attendance]
      summary: Student scans the session QR (student)
      description: |
        The record is bound to the authenticated student. A second scan is idempotent: it returns
        `duplicate: true` and never creates another record. A student is `late` when scanning more than
        15 minutes after the session was opened.

        Rejected scans return `accepted: false` with a `reason`:
        `invalid_token`, `expired_token`, `session_closed` (HTTP 400), `not_enrolled` (HTTP 403),
        `manually_recorded` (HTTP 409, the lecturer/TA already set this student's status manually,
        so a scan cannot change it) or `too_many_attempts` (HTTP 429).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: The token read from the QR code.
      responses:
        '200':
          description: Scan accepted (or duplicate of an earlier accepted scan)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanAccepted'
        '400':
          description: Scan rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanRejected'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not a student, or the student is not enrolled in this session
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ScanRejected'
                  - $ref: '#/components/schemas/Error'
        '409':
          description: The record was set manually by staff; the scan does not change it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanRejected'
              example:
                accepted: false
                reason: manually_recorded
                attendance_status: excused
        '429':
          description: Too many scan attempts (limit 10 per minute per account). Wait a minute and retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanRejected'
              example:
                accepted: false
                reason: too_many_attempts
  /attendance/me:
    get:
      tags: [Attendance]
      summary: The student's own attendance history (student)
      responses:
        '200':
          description: Own records only
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MyAttendanceRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Missing, invalid or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The role is not allowed, or the record belongs to someone else
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Role:
      type: string
      enum: [student, lecturer, ta, admin, auditor]
...
    AttendanceStatus:
      type: string
      enum: [present, absent, late, excused]
...
    ScanAccepted:
      type: object
      properties:
        accepted:
          type: boolean
          example: true
        duplicate:
          type: boolean
          description: true when this student was already recorded (no new record is created)
        attendance_status:
          $ref: '#/components/schemas/AttendanceStatus'
        minutes_late:
          type: integer
        attendance_timestamp:
          type: string
          format: date-time
          description: Only present when duplicate is false
    ScanRejected:
      type: object
      properties:
        accepted:
          type: boolean
          example: false
        reason:
          type: string
          example: expired_token
          description: invalid_token, expired_token, session_closed, not_enrolled, manually_recorded, too_many_attempts, or a validation message
    MyAttendanceRecord:
      type: object
      properties:
        attendance_id:
          type: integer
        course_code:
          type: string
        course_name:
          type: string
        session_date:
          type: string
          format: date
        attendance_status:
          $ref: '#/components/schemas/AttendanceStatus'
        minutes_late:
          type: integer
        attendance_timestamp:
          type: string
          format: date-time
          nullable: true
...